+++ pkg/cmd/cdev/project.go
@@ -2,7 +2,9 @@ package cdev
 
 import (
 	"github.com/apex/log"
+	"github.com/shalb/cluster.dev/pkg/config"
 	"github.com/shalb/cluster.dev/pkg/project"
+	"github.com/shalb/cluster.dev/pkg/project/ui"
 	"github.com/spf13/cobra"
 )
 
@@ -15,6 +17,7 @@ var projectCmd = &cobra.Command{
 func init() {
 	rootCmd.AddCommand(projectCmd)
 	projectCmd.AddCommand(projectLs)
+	projectCmd.AddCommand(projectCreate)
 }
 
 // projectsCmd represents the plan command
@@ -29,5 +32,20 @@ var projectLs = &cobra.Command{
 		}
 		log.Info("Project info:")
 		p.PrintInfo()
+	},
+}
+
+// projectsCmd represents the plan command
+var projectCreate = &cobra.Command{
+	Use:   "create",
+	Short: "Generate new project from template in curent dir. Directory must be empty",
+	Run: func(cmd *cobra.Command, args []string) {
+		if project.ProjectsFilesExists() {
+			log.Fatalf("project creating: some project's data (yaml files) found in current directory, use command in empty dir")
+		}
+		err := ui.CreteProject(config.Global.WorkingDir)
+		if err != nil {
+			log.Fatalf("Create project: %v", err.Error())
+		}
 	},
 }
